refector(api): move API_VERSION const out of ActixWeb implementation
We are going to remove the ActixWeb implementation after enabling the
new Axum implementaion. WE are moving generic things in ActixWeb
implementations that are not going to be removed.
