+++ src/routes/about.rs
@@ -2,8 +2,8 @@ use actix_web::http::StatusCode;
 use actix_web::{web, HttpResponse, Responder};
 
 use crate::errors::ServiceResult;
-use crate::routes::API_VERSION;
 use crate::services::about::{index_page, license_page};
+use crate::web::api::API_VERSION;
 
 pub fn init(cfg: &mut web::ServiceConfig) {
     cfg.service(
