+++ src/routes/mod.rs
@@ -9,8 +9,6 @@ pub mod tag;
 pub mod torrent;
 pub mod user;
 
-pub const API_VERSION: &str = "v1";
-
 pub fn init(cfg: &mut web::ServiceConfig) {
     user::init(cfg);
     torrent::init(cfg);
