+++ src/routes/tag.rs
@@ -5,7 +5,7 @@ use crate::common::WebAppData;
 use crate::errors::ServiceResult;
 use crate::models::response::OkResponse;
 use crate::models::torrent_tag::TagId;
-use crate::routes::API_VERSION;
+use crate::web::api::API_VERSION;
 
 pub fn init(cfg: &mut web::ServiceConfig) {
     cfg.service(
